refactor: src folders
