+++ .env.example
@@ -17,5 +17,5 @@ PGADMIN_DEFAULT_PASSWORD=admin
 
 ACCESS_TOKEN_EXPIRATION_TIME=60s
 ACCESS_TOKEN_SECRET_KEY=secret
-REFRESH_TOKEN_EXPIRATION_TIME=80s
+REFRESH_TOKEN_EXPIRATION_TIME=800s
 REFRESH_TOKEN_SECRET_KEY=secret refresh
